- bottom_sheet вынесен в BottomSheetDialogFragment
