+++ app/src/main/java/com/example/playlistmaker/player/di/ViewModelModule.kt
@@ -1,11 +1,20 @@
 package com.example.playlistmaker.player.di
 
+import com.example.playlistmaker.player.presentation.PlayerBottomSheetViewModel
 import com.example.playlistmaker.player.presentation.PlayerViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val playerViewModelModule = module {
     viewModel {
-        PlayerViewModel(playerInteractor = get(), favoritesInteractor = get(), playListsInteractor = get())
+        PlayerViewModel(
+            playerInteractor = get(),
+            favoritesInteractor = get()
+        )
+    }
+    viewModel {
+        PlayerBottomSheetViewModel(
+            playListsInteractor = get()
+        )
     }
 }
