+++ app/src/main/java/com/example/playlistmaker/player/presentation/PlayerFragment.kt
@@ -5,9 +5,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.LinearLayout
-import android.widget.Toast
-import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
@@ -16,10 +13,7 @@ import com.example.playlistmaker.R
 import com.example.playlistmaker.common.TRACK
 import com.example.playlistmaker.common.models.Track
 import com.example.playlistmaker.databinding.FragmentPlayerBinding
-import com.example.playlistmaker.medialibrary.domain.models.PlayList
-import com.example.playlistmaker.player.presentation.models.PlayListsState
 import com.example.playlistmaker.player.presentation.models.PlayerState
-import com.google.android.material.bottomsheet.BottomSheetBehavior
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import java.text.SimpleDateFormat
 import java.util.Locale
@@ -30,13 +24,7 @@ class PlayerFragment: Fragment() {
 
     private val viewModel by viewModel<PlayerViewModel>()
 
-    private val playListsAdapter = PlayListsAdapter {
-        addTrackToPlayList(it)
-    }
-
     private lateinit var track: Track
-
-    private lateinit var bottomSheetBehavior: BottomSheetBehavior<LinearLayout>
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -50,21 +38,15 @@ class PlayerFragment: Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.playListsRV.adapter = playListsAdapter
-
         showTrack(track)
 
         viewModel.preparePlayer(track.previewUrl)
 
         viewModel.isFavorite(track.trackId)
 
-        initBottomSheet()
-
         initObserveViewModel()
 
         initOnClickListeners()
-
-        initOnBackPressedCallback()
 
     }
 
@@ -78,41 +60,18 @@ class PlayerFragment: Fragment() {
         viewModel.pausePlayer()
     }
 
-    override fun onResume() {
-        super.onResume()
-        viewModel.getPlayLists()
-    }
-
-    private fun initOnBackPressedCallback() {
-        val callback = object : OnBackPressedCallback(true) {
-            override fun handleOnBackPressed() {
-                if (bottomSheetBehavior.state != BottomSheetBehavior.STATE_HIDDEN) {
-                    bottomSheetBehavior.state = BottomSheetBehavior.STATE_HIDDEN
-                } else {
-                    findNavController().popBackStack()
-                }
-            }
-        }
-        requireActivity().onBackPressedDispatcher.addCallback(requireActivity(), callback)
-    }
-
     private fun initOnClickListeners() {
         binding.toolbar.setNavigationOnClickListener {
             findNavController().popBackStack()
         }
         binding.playlistButton.setOnClickListener {
-            bottomSheetBehavior.state = BottomSheetBehavior.STATE_EXPANDED
+            PlayerBottomSheetFragment.newInstance(track).show(childFragmentManager,PlayerBottomSheetFragment.TAG)
         }
         binding.favoriteButton.setOnClickListener {
             viewModel.onFavoriteClicked(track)
         }
         binding.playButton.setOnClickListener {
             viewModel.playbackControl()
-        }
-        binding.addPlaylistButton.setOnClickListener {
-            findNavController().navigate(
-                R.id.action_playerFragment_to_addPlayListFragment
-            )
         }
     }
 
@@ -126,53 +85,6 @@ class PlayerFragment: Fragment() {
         viewModel.observeFavoriteState().observe(viewLifecycleOwner) {
             render(it)
         }
-        viewModel.observePlayListsState().observe(viewLifecycleOwner) {
-            when (it) {
-                is PlayListsState.Empty -> binding.playListsRV.visibility = View.GONE
-                is PlayListsState.PlayLists -> {
-                    playListsAdapter.playLists = it.playLists
-                    binding.playListsRV.visibility = View.VISIBLE
-                }
-                is PlayListsState.AddTrackToPlayListResult -> {
-                    if (it.isAdded) {
-                        showToast(getString(R.string.added_to_playlist, it.playListName))
-                    } else {
-                        showToast(getString(R.string.already_added_to_playlist, it.playListName))
-                    }
-                    bottomSheetBehavior.state = BottomSheetBehavior.STATE_HIDDEN
-                }
-            }
-        }
-    }
-
-    private fun initBottomSheet() {
-        val bottomSheetContainer = binding.playerBottomSheet
-        bottomSheetBehavior = BottomSheetBehavior.from(bottomSheetContainer).apply {
-            state = BottomSheetBehavior.STATE_HIDDEN
-        }
-        bottomSheetBehavior.addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
-            override fun onStateChanged(bottomSheet: View, newState: Int) {
-                when (newState) {
-                    BottomSheetBehavior.STATE_HIDDEN -> {
-                        binding.overlay.visibility = View.GONE
-                    }
-                    else -> {
-                        binding.overlay.visibility = View.VISIBLE
-                    }
-                }
-            }
-            override fun onSlide(bottomSheet: View, slideOffset: Float) {}
-        })
-    }
-
-    private fun addTrackToPlayList(playList: PlayList) {
-        if (viewModel.clickDebounce()) {
-            viewModel.addTrackToPlayList(track, playList)
-        }
-    }
-
-    private fun showToast(additionalMessage: String) {
-        Toast.makeText(requireContext(), additionalMessage, Toast.LENGTH_LONG).show()
     }
 
     private fun render(state: PlayerState) {
