+++ app/src/main/res/navigation/navigation_graph.xml
@@ -57,4 +57,13 @@
             android:id="@+id/action_playerFragment_to_addPlayListFragment"
             app:destination="@id/addPlayListFragment" />
     </fragment>
+    <dialog
+        android:id="@+id/bottomSheetFragment"
+        android:name="com.example.playlistmaker.player.presentation.PlayerBottomSheetFragment"
+        android:label="fragment_bottom_sheet"
+        tools:layout="@layout/fragment_bottom_sheet" >
+        <action
+            android:id="@+id/action_bottomSheetFragment_to_addPlayListFragment"
+            app:destination="@id/addPlayListFragment" />
+    </dialog>
 </navigation>
